cleanup(philo_utils.c, philo.c): cleaned the routines function and created a function to print the right message to avoid repetition in the thread's routine
