+++ philo/srcs/exec/philo.c
@@ -6,100 +6,56 @@
 void* routine_odd(void *arg)
 {
     t_philo *philo = (t_philo *)arg;
-    struct timeval tv;
-
-    usleep(5000);
-//moitie wait la moitie de manger pour creer un decalage
+    //struct timeval tv;
+    //TODO mettre le t_eat /2 directement dans la struct pour eviter calcul redondant
+    usleep(250);//moitie wait la moitie de manger pour creer un decalage
     while(!philo->set->death)
     {
-        gettimeofday(&tv, NULL);
-        philo->set->time_passed = (tv.tv_sec -  philo->set->subunit) * 1000000 + (tv.tv_usec - philo->set->subusec);
-        pthread_mutex_lock(&philo->set->print_mutex);
-        printf("%ld %ld is thinking\n",philo->set->time_passed / 1000, philo->id);
-        pthread_mutex_unlock(&philo->set->print_mutex);
+        print_msg_routine(philo, IS_THINKING);
         pthread_mutex_lock(&philo->right->mutex);
         philo->right->available = false;
-        gettimeofday(&tv, NULL);
-        philo->set->time_passed = (tv.tv_sec -  philo->set->subunit) * 1000000 + (tv.tv_usec - philo->set->subusec);
-        pthread_mutex_lock(&philo->set->print_mutex);
-        printf("%ld %ld has taken a fork\n",philo->set->time_passed / 1000, philo->id);
-        pthread_mutex_unlock(&philo->set->print_mutex);
+        print_msg_routine(philo, IS_TAKING_FORK);
         pthread_mutex_lock(&philo->left->mutex);
         philo->left->available = false;
-        gettimeofday(&tv, NULL);
-        philo->set->time_passed = (tv.tv_sec -  philo->set->subunit) * 1000000 + (tv.tv_usec - philo->set->subusec);
-        pthread_mutex_lock(&philo->set->print_mutex);
-        printf("%ld %ld has taken a fork\n",philo->set->time_passed / 1000, philo->id);
-        pthread_mutex_unlock(&philo->set->print_mutex);
-        philo->set->time_passed = (tv.tv_sec -  philo->set->subunit) * 1000000 + (tv.tv_usec - philo->set->subusec);
-        pthread_mutex_lock(&philo->set->print_mutex);
-        printf("%ld %ld is eating\n",philo->set->time_passed / 1000, philo->id);
-        pthread_mutex_unlock(&philo->set->print_mutex);
+        print_msg_routine(philo, IS_TAKING_FORK);
+        print_msg_routine(philo, IS_EATING);
         usleep(philo->set->t_eat);
         philo->left->available = true;
         pthread_mutex_unlock(&philo->left->mutex);
         philo->right->available = true;
         pthread_mutex_unlock(&philo->right->mutex);
         //reset time eat
-        gettimeofday(&tv, NULL);
-        philo->set->time_passed = (tv.tv_sec -  philo->set->subunit) * 1000000 + (tv.tv_usec - philo->set->subusec);
-        pthread_mutex_lock(&philo->set->print_mutex);
-        printf("%ld %ld is sleeping\n",philo->set->time_passed / 1000, philo->id);
-        pthread_mutex_unlock(&philo->set->print_mutex);
-        usleep(philo->set->t_sleep);
-            
-            //TODO else if didnt ate in time
-            //kill
-            // gettimeofday(&tv, NULL);
-            // philo->set->time_passed = (tv.tv_sec -  philo->set->subunit) * 1000000 + (tv.tv_usec - philo->set->subusec);
-            // printf("[%ld] %ld died\n",philo->set->time_passed / 1000, philo->id);
+        print_msg_routine(philo, IS_SLEEPING);
+        usleep(philo->set->t_sleep);   
+        //TODO else if didnt ate in time
+        //kill
+        // gettimeofday(&tv, NULL);
+        // philo->set->time_passed = (tv.tv_sec -  philo->set->subunit) * 1000000 + (tv.tv_usec - philo->set->subusec);
+        // printf("[%ld] %ld died\n",philo->set->time_passed / 1000, philo->id);
     }
-    // pthread_mutex_lock(&philo->set->print_mutex);
-    // pthread_mutex_unlock(&philo->set->print_mutex);
     return(0);
 }
 
 void* routine_even(void *arg)
 {
     t_philo *philo = (t_philo *)arg;
-    struct timeval tv;
-//moitie wait la moitie de manger pour creer un decalage
     while(!philo->set->death)
     {
-        gettimeofday(&tv, NULL);
-        philo->set->time_passed = (tv.tv_sec -  philo->set->subunit) * 1000000 + (tv.tv_usec - philo->set->subusec);
-        pthread_mutex_lock(&philo->set->print_mutex);
-        printf("%ld %ld is thinking\n",philo->set->time_passed / 1000, philo->id);
-        pthread_mutex_unlock(&philo->set->print_mutex);
+        print_msg_routine(philo, IS_THINKING);
         pthread_mutex_lock(&philo->left->mutex);
         philo->left->available = false;
-        gettimeofday(&tv, NULL);
-        philo->set->time_passed = (tv.tv_sec -  philo->set->subunit) * 1000000 + (tv.tv_usec - philo->set->subusec);
-        pthread_mutex_lock(&philo->set->print_mutex);
-        printf("%ld %ld has taken a fork\n",philo->set->time_passed / 1000, philo->id);
-        pthread_mutex_unlock(&philo->set->print_mutex);
+        print_msg_routine(philo, IS_TAKING_FORK);
         pthread_mutex_lock(&philo->right->mutex);
         philo->right->available = false;
-        gettimeofday(&tv, NULL);
-        philo->set->time_passed = (tv.tv_sec -  philo->set->subunit) * 1000000 + (tv.tv_usec - philo->set->subusec);
-        pthread_mutex_lock(&philo->set->print_mutex);
-        printf("%ld %ld has taken a fork\n",philo->set->time_passed / 1000, philo->id);
-        pthread_mutex_unlock(&philo->set->print_mutex);
-        philo->set->time_passed = (tv.tv_sec -  philo->set->subunit) * 1000000 + (tv.tv_usec - philo->set->subusec);
-        pthread_mutex_lock(&philo->set->print_mutex);
-        printf("%ld %ld is eating\n",philo->set->time_passed / 1000, philo->id);
-        pthread_mutex_unlock(&philo->set->print_mutex);
+        print_msg_routine(philo, IS_TAKING_FORK);
+        print_msg_routine(philo, IS_EATING);
         usleep(philo->set->t_eat);
         philo->left->available = true;
         pthread_mutex_unlock(&philo->left->mutex);
         philo->right->available = true;
         pthread_mutex_unlock(&philo->right->mutex);
         //reset time eat
-        gettimeofday(&tv, NULL);
-        philo->set->time_passed = (tv.tv_sec -  philo->set->subunit) * 1000000 + (tv.tv_usec - philo->set->subusec);
-        pthread_mutex_lock(&philo->set->print_mutex);
-        printf("%ld %ld is sleeping\n",philo->set->time_passed / 1000, philo->id);
-        pthread_mutex_unlock(&philo->set->print_mutex);
+        print_msg_routine(philo, IS_SLEEPING);
         usleep(philo->set->t_sleep);    
             //TODO else if didnt ate in time
             //kill
